Added Hasher class
