+++ Bot.Integration.Git.Tests/AddFileCommandTests.cs
@@ -1,4 +1,5 @@
 ﻿using Bot.Integration.Git.GitCommands.AddFile;
+using Bot.Integration.Git.Tests;
 using System.Security.Cryptography;
 using System.Text;
 
@@ -28,31 +29,15 @@ public class AddFileCommandTests
     {
         var originalFilepath = $"Fixtures/{filename}";
         await using var stream = File.OpenRead(originalFilepath);
-        var expectedHash = GetHashString(originalFilepath);
+        var expectedHash = Hasher.GetHashString(originalFilepath);
         var repositoryFilepath = Path.Combine(REPOSITORY_PATH, filename);
         var request = new AddFileCommand(stream, repositoryFilepath);
         var handler = new AddFileCommandHandler();
 
         await handler.Handle(request, default);
 
-        var actualhash = GetHashString(repositoryFilepath);      
+        var actualhash = Hasher.GetHashString(repositoryFilepath);      
         Assert.That(actualhash, Is.EqualTo(expectedHash));
 
-    }
-    private static string GetHashString(string filepath)
-    {
-        var sha256 = SHA256.Create();
-        using var fileStream = File.OpenRead(filepath);
-        return ByteArrayToHexString(sha256.ComputeHash(fileStream));
-    }
-
-    private static string ByteArrayToHexString(byte[] bytes)
-    {
-        string output = string.Empty;
-        foreach (var @byte in bytes)
-        {
-            output += $"{@byte:x2}";
-        }
-        return output;
-    }
+    }    
 }
